Make same as development

+++ src/main/java/cz/fi/muni/pv168/easyfood/ui/action/DeleteAction.java
@@ -5,9 +5,16 @@ import cz.fi.muni.pv168.easyfood.ui.Icons;
 import cz.fi.muni.pv168.easyfood.ui.tab.TabContainer;
 
 import javax.swing.AbstractAction;
+import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+import javax.swing.JPanel;
 import javax.swing.KeyStroke;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
+
+import static javax.swing.JOptionPane.OK_OPTION;
+import static javax.swing.JOptionPane.QUESTION_MESSAGE;
+import static javax.swing.JOptionPane.YES_NO_OPTION;
 
 public final class DeleteAction extends AbstractAction {
     private final TabContainer tabContainer;
@@ -23,7 +30,20 @@ public final class DeleteAction extends AbstractAction {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        tabContainer.getSelectedTab().delete();
+        var table = tabContainer.getSelectedTab().getTable();
+        int selectedRows = table.getSelectedRows().length;
+        if (selectedRows < 1) {
+            throw new IllegalStateException("Invalid selected rows count (must be 1): " + selectedRows);
+        }
+
+        JPanel jPanel = new JPanel();
+        jPanel.add(new JLabel("Delete " + selectedRows + ((selectedRows == 1) ? " row?" : " rows?")));
+
+        int result = JOptionPane.showConfirmDialog(table, jPanel, "Delete",
+                YES_NO_OPTION, QUESTION_MESSAGE, null);
+        if (result == OK_OPTION) {
+            tabContainer.getSelectedTab().delete();
+        }
     }
 }
